+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# v 0.3.2
+
+# New features
+
+* added new config param `oracle_discovery_skip_errors_regex` to count as errors only those which we consider errors to check.
+* added discover_stats metrics, `connect_errors_skipped`
+
+## Breaking changes.
+
+* changed discover_stats metrics names:
+  - from `disconnected` to `undiscovered`
+  - from `disconnected_sid_names` to `undiscovered_sid_names`
+
+# v 0.3.1 (2023-03-13)
+
+## New features
+
+## Fixes
+
+fixed errconnect_sid_names from the discover_stats statistics
+
 # v 0.1.0 (2022-10-22)
 
 ## New features